# internal/app/rest_api_v2/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "`NOTE!` This REST API HTTP endpoint is located directly on the `Hoster` node.\u003cbr\u003e\u003cbr\u003eThe API should ideally be integrated into another system (e.g. a user-accessible back-end server), and not interacted with directly.\u003cbr\u003e\u003cbr\u003ePlease, take an extra care with the things you execute here, because some of them may be disruptive or non-revertible (e.g. vm destroy, snapshot rollback, host reboot, etc).",
        "title": "Hoster Node REST API Docs",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "basePath": "/api/v2",
    "paths": {
...
        "/scheduler/jobs": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Get the list of scheduled, active and past jobs.\u003cbr\u003e`AUTH`: Only REST user is allowed.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scheduler"
                ],
                "summary": "Get the list of scheduled, active and past jobs.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/SchedulerUtils.Job"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "SchedulerUtils.Job": {
            "type": "object",
            "properties": {
                "job_done": {
                    "type": "boolean"
                },
                "job_done_logged": {
                    "type": "boolean"
                },
                "job_error": {
                    "type": "string"
                },
                "job_failed": {
                    "type": "boolean"
                },
                "job_failed_logged": {
                    "type": "boolean"
                },
                "job_id": {
                    "type": "string"
                },
                "job_in_progress": {
                    "type": "boolean"
                },
                "job_next": {
                    "type": "boolean"
                },
                "job_type": {
                    "type": "string"
                },
                "replication": {
                    "$ref": "#/definitions/SchedulerUtils.ReplicationJob"
                },
                "res_type": {
                    "type": "string"
                },
                "snapshot": {
                    "$ref": "#/definitions/SchedulerUtils.SnapshotJob"
                },
                "time_added": {
                    "type": "integer"
                },
                "time_finished": {
                    "type": "integer"
                }
            }
        },
        "SchedulerUtils.ReplicationJob": {
            "type": "object",
            "properties": {
                "done_snaps": {
                    "type": "integer"
                },
                "progress_bytes_done": {
                    "type": "integer"
                },
                "progress_bytes_total": {
                    "type": "integer"
                },
                "res_name": {
                    "type": "string"
                },
                "scripts_remove": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "scripts_replicate": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "speed_limit": {
                    "type": "integer"
                },
                "ssh_endpoint": {
                    "type": "string"
                },
                "ssh_key": {
                    "type": "string"
                },
                "ssh_port": {
                    "type": "integer"
                },
                "total_snaps": {
                    "type": "integer"
                },
                "zfs_dataset": {
                    "type": "string"
                }
            }
        },
        "SchedulerUtils.SnapshotJob": {
            "type": "object",
            "properties": {
                "res_name": {
                    "type": "string"
                },
                "snapshot_name": {
                    "description": "only used in the snapshot destroy jobs",
                    "type": "string"
                },
                "snapshot_type": {
                    "type": "string"
                },
                "snapshots_to_keep": {
                    "type": "integer"
                },
                "take_immediately": {
                    "type": "boolean"
                },
                "zfs_dataset": {
                    "type": "string"
                }
            }
        },
...
        "handlers.SwaggerError": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}